+++ app/src/main/java/woowacourse/movie/domain/repository/ScreenRepository.kt
@@ -3,7 +3,6 @@ package woowacourse.movie.domain.repository
 import woowacourse.movie.domain.model.Screen
 import woowacourse.movie.domain.model.ScreenView
 import woowacourse.movie.domain.model.SeatBoard
-import woowacourse.movie.domain.model.TheaterCount
 
 interface ScreenRepository {
     fun load(): List<ScreenView>
@@ -14,8 +13,4 @@ interface ScreenRepository {
         theaterId: Int,
         movieId: Int,
     ): Result<Screen>
-
-    fun findTheaterCount(id: Int): Result<List<TheaterCount>>
-
-    fun findTheaterNameById(theaterId: Int): Result<String>
 }
