+++ app/src/main/java/woowacourse/movie/domain/repository/TheaterRepository.kt
@@ -1,0 +1,9 @@
+package woowacourse.movie.domain.repository
+
+import woowacourse.movie.domain.model.TheaterCount
+
+interface TheaterRepository {
+    fun findTheaterCount(id: Int): Result<List<TheaterCount>>
+
+    fun findTheaterNameById(theaterId: Int): Result<String>
+}
