+++ app/src/main/java/woowacourse/movie/presentation/ui/reservation/ReservationPresenter.kt
@@ -1,17 +1,17 @@
 package woowacourse.movie.presentation.ui.reservation
 
 import woowacourse.movie.domain.repository.ReservationRepository
-import woowacourse.movie.domain.repository.ScreenRepository
+import woowacourse.movie.domain.repository.TheaterRepository
 
 class ReservationPresenter(
     private val view: ReservationContract.View,
-    private val repository: ReservationRepository,
-    private val theaterRepository: ScreenRepository,
+    private val reservationRepository: ReservationRepository,
+    private val theaterRepository: TheaterRepository,
 ) : ReservationContract.Presenter {
     private var uiModel: ReservationUiModel = ReservationUiModel()
 
     override fun loadReservation(id: Int) {
-        repository.findByReservationId(id).onSuccess { reservation ->
+        reservationRepository.findByReservationId(id).onSuccess { reservation ->
             theaterRepository.findTheaterNameById(reservation.theaterId).onSuccess { theaterName ->
                 view.showReservation(reservation, theaterName)
                 uiModel = uiModel.copy(theaterName = theaterName, reservation = reservation)
